Refactored report.js file name to better reflect its purpose.
