+++ .gitignore
@@ -1,1 +1,1 @@
-.login_credentials
+login_credentials.json
